Re-arrange some files, add some brief documentation
