+++ shm/README.md
@@ -1,0 +1,6 @@
+# Shared Memory
+
+`SimSharedMemory.py` defines a class that the server uses to manage shared memory.
+
+`SimSharedMemoryClient.py` defines a class that can be used by client applications to read from the shared memory.
+It purposely does not contain any GeoScenario dependencies.
